requirements.txt
typer
tqdm
pyyaml
requests
validators
beautifulsoup4
pyinstaller